Crate exec

exec.rs

The module defines basics for asynchronous task and its execution.

An asynchronous task is a function whose operation would be completed in future.

 ------------------------
| Sender -----> Receiver |
 ------------------------
    Operation State

The above diagram shows the model for async task.

Structs

InlineScheduler

Scheduler that schedules task on current thread.

NeverStopCallback

NeverStopToken

A stop token, that never stops.

Traits

OperationState

Represents executable task.

Receiver

This models the part of task which contains callback when task completes.

Scheduler

Represents handle to execution context.

Sender

This models the part of task which represents the work upon completion of which, receiver’s callback should be called.

SenderExt

StopCallback

StopSource

StopToken

Functions

just

Returns a sender, that completes on current context with given value.

just_cancelled

Returns a sender, that completes on current context with indicating cancellation.

just_error

Returns a sender, that completes on current context with given error.

let_value

sync_wait

Perform a blocking wait on sender and return result of its computation.

then

Returns a sender which completes with value continuation(InternalSender::Value).