exec
Trait Receiver

pub trait Receiver {
    type Value;
    type Error;

    // Required methods
    fn set_value(self, value: Self::Value);
    fn set_error(self, error: Self::Error);
    fn set_cancelled(self);
}

This models the part of task which contains callback when task completes.

A task can be completed with Value, Error or cancelled.

Required Associated Types

type Value

Type of value with which task can be completed.

type Error

Type of error with which task can be completed.

Required Methods

fn set_value(self, value: Self::Value)

Callback for task, when it completes with value.

fn set_error(self, error: Self::Error)

Callback for task, when it completes with error.

fn set_cancelled(self)

Callback for task, when it gets cancelled.