exec
Trait StopToken

pub trait StopToken {
    type CallbackHandle: StopCallback;

    // Required methods
    fn stop_requested(&self) -> bool;
    fn register_callback<F>(&mut self, callback: F) -> Self::CallbackHandle
       where F: FnMut();
}

Required Associated Types

type CallbackHandle: StopCallback

Required Methods

fn stop_requested(&self) -> bool

Returns true if stop token has stop requested.

fn register_callback<F>(&mut self, callback: F) -> Self::CallbackHandle
where F: FnMut(),

Creates a new callback that would be called when stop would be requested.

Callback would be called on same thread which did request stop. If token would already have stop_requested() == true at time of creation of callback, then callback would be called on thread creating stop callback.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.